pyproject.toml
[tool.poetry]
name = "PeterThePedant_bot"
version = "0.1.0"
description = "A Grammar Nazi here to always tell you correct definitions of words."
authors = ["Femi Ayo-Tubosun <femiayotubosun@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.8.1"
python-dotenv = "^0.19.2"
black = "^21.12b0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
